Yolo_OCR/Flask/requirements.txt
Flask
opencv-python 
numpy 
requests 
Werkzeug 
ultralytics 
google-cloud-vision